refactor(format): split kdbx4 into modules
